+++ spec/unit/mutant/cli_spec.rb
@@ -126,7 +126,6 @@ Environment:
 
 Options:
         --expected-coverage COVERAGE Fail unless COVERAGE is not reached exactly, parsed via Rational()
-        --score COVERAGE             Fail unless COVERAGE is not reached exactly [deprecated]
         --use INTEGRATION            Use INTEGRATION to kill mutations
         --ignore-subject PATTERN     Ignore subjects that match PATTERN
         --fail-fast                  Fail fast
@@ -218,16 +217,6 @@ Options:
         it 'configures expected coverage' do
           expect(subject.config.expected_coverage).to eql(Rational(1, 10))
         end
-      end
-    end
-
-    context 'with score flag' do
-      let(:flags) { %w[--score 50.0] }
-
-      it_should_behave_like 'a cli parser'
-
-      it 'configures expected coverage' do
-        expect(subject.config.expected_coverage).to eql(Rational(1, 2))
       end
     end