+++ Changelog.md
@@ -2,6 +2,7 @@
 
 * Drop support for ruby < 2.1
 * Remove broken `--code` option
+* Remove deprecated `--score` option
 
 # v0.7.9 2015-05-30